Promote CachedDriveInfo to top-level internal type with injectable acquire delegate
CachedDriveInfo is extracted from its nested position inside LockDirectory
into its own file. DetectLockSupport gains an overload that accepts a
Func<LockOptions, ILockHandle> delegate in place of the direct FileLock.Acquire
call, making the four lock-probe tests (exclusive, shared, exclusive-blocks-shared,
shared-blocks-exclusive) unit-testable without real filesystem operations.
The existing public overload forwards to FileLock.Acquire.
